Update SumApp type int to long

+++ SumApp/SumApp.cs
@@ -8,9 +8,9 @@ namespace SumApp
 {
     public static class SumApp
     {
-        public static int Sum(int[] arr)
+        public static long Sum(int[] arr)
         {
-            int sum = 0;
+            long sum = 0;
 
             for (int i = 0; i < arr.Length; i++)
             {
@@ -22,7 +22,7 @@ namespace SumApp
 
         public static double Average(int[] arr)
         {
-            int sum = 0;
+            double sum = 0;
 
             for (int i = 0; i < arr.Length; i++)
             {
